+++ R/files.R
@@ -5,14 +5,12 @@
 #' @param ... Sink for unused
 #'
 #' @return List of paths to the respective files
-#' @export
 #' @md
-#' @family UI
 #' @importFrom tools file_path_sans_ext file_ext
 #'
 #' @examples
-#' path <- system.file("exampleData/01-Iris.pdf", package = "excelPlot")
-#' epFiles(path, commit = "HEAD", page = 1)
+#' path <- system.file("exampleData/01-Iris.pdf", package = "plotExcel")
+#' plotExcel:::epFiles(path, commit = "HEAD", page = 1)
 epFiles <- function(path, commit = "HEAD", page = 1, resolution = 100, xmin = 0, xmax = 100, ymin = 0, ymax = 100, ...) {
 
   path <- normalizePath(path, mustWork = TRUE, winslash = .Platform$file.sep)